+++ Sources/AutoSDK/AutoSDKVersion.m
@@ -1,4 +1,4 @@
 #import "include/AutoSDK.h"
 
-double AutoSDKVersionNumber = 1.17;
-const unsigned char AutoSDKVersionString[] = "1.17.0";
+double AutoSDKVersionNumber = 1.18;
+const unsigned char AutoSDKVersionString[] = "1.18.0";